HasChamber trait to share chamber funcs
